app/chat/requirements.txt
flask
line-bot-sdk
openai==0.28
python-dotenv